Revert "Move mk3, Scale mk3  and Rotation m3 (#2899)"
This reverts commit cb0ca3bedbacf20fb5b370ccc82501874c302c5f.
