+++ docs/nodes/transforms/scale_mk2.rst
@@ -13,8 +13,7 @@ All inputs are vectorized and they will accept single or multiple values.
 There is three inputs:
 
 - **Vertices**
-- **Centers**
-- **Scale**
+- **Center**
 - **Factor**
 
 Parameters
@@ -24,15 +23,13 @@ All parameters except **Vertices** has a default value. **Factor** can be given 
 
 
 +----------------+---------------+-----------------+----------------------------------------------------+
-| Param          | Type          | Default         | Description                                        |
+| Param          | Type          | Default         | Description                                        |  
 +================+===============+=================+====================================================+
-| **Vertices**   | Vector        | none            | vertices to scale                                  |
+| **Vertices**   | Vertices      | none            | vertices to scale                                  | 
 +----------------+---------------+-----------------+----------------------------------------------------+
-| **Centers**    | Vector        | (0.0, 0.0, 0.0) | point from which the scaling will be done          |
+| **Center**     | Vertices      | (0.0, 0.0, 0.0) | point from which the scaling will be done          |
 +----------------+---------------+-----------------+----------------------------------------------------+
-| **Scale**      | Vector        | (1.0, 1.0, 1.0) | Axis scaling                                       |
-+----------------+---------------+-----------------+----------------------------------------------------+
-| **Multiplier** | Float         | 1.0             | Uniform multiplier factor                          |
+| **Factor**     | Float         | 1.0             | factor of scaling                                  |
 +----------------+---------------+-----------------+----------------------------------------------------+
 
 Outputs
@@ -44,14 +41,7 @@ If you generate more outputs than inputs were given, then is probably that you n
 Example of usage
 ----------------
 
-Creating a spiral from a Circle:
+.. image:: https://cloud.githubusercontent.com/assets/5990821/4220292/73e9430e-3900-11e4-949f-02749baa7751.png
+  :alt: ScaleDemo1.PNG
 
-.. image:: https://raw.githubusercontent.com/vicdoval/sverchok/docs_images/images_for_docs/transforms/scale/scale_vectors_blender_sverchok_example_1.png
-
-The node will match different data structures, in this case this mechanism is used to create surface is generated from a line:
-
-.. image:: https://raw.githubusercontent.com/vicdoval/sverchok/docs_images/images_for_docs/transforms/scale/scale_vectors_blender_sverchok_example_2.png
-
-Deforming an icosphere into a distorted ovoid:
-
-.. image:: https://raw.githubusercontent.com/vicdoval/sverchok/docs_images/images_for_docs/transforms/scale/scale_vectors_blender_sverchok_example_3.png
+In this example we use scale to convert a simple circle into a kind of parabola.
