+++ docs/nodes/transforms/transforms_index.rst
@@ -5,9 +5,9 @@ Transforms
 .. toctree::
    :maxdepth: 2
 
-   move_mk3
-   scale_mk3
-   rotate_mk3
+   move_mk2
+   scale_mk2
+   rotation_mk2
    apply
    barycentric_transform
    bend_along_path
