+++ json_examples/Design/Adaptive_Voronoi_panel.json
@@ -10,7 +10,7 @@
     "Map Range": "Frame.001",
     "Map Range.001": "Frame",
     "Mesh Expression": "Frame",
-    "Move.001": "Frame.003",
+    "Move": "Frame.003",
     "Note": "Frame.002",
     "Note.001": "Frame.002",
     "Origins": "Frame.001",
@@ -38,9 +38,7 @@
       ],
       "params": {
         "float_": 2.380000114440918,
-        "float_draft_": 0.0,
-        "int_": 0,
-        "int_draft_": 0
+        "int_": 0
       },
       "width": 140.0
     },
@@ -55,9 +53,7 @@
       ],
       "params": {
         "float_": 0.6699999570846558,
-        "float_draft_": 0.0,
-        "int_": 0,
-        "int_draft_": 0
+        "int_": 0
       },
       "width": 140.0
     },
@@ -111,7 +107,7 @@
     },
     "Frame": {
       "bl_idname": "NodeFrame",
-      "height": 319.0199279785156,
+      "height": 317.0199279785156,
       "hide": false,
       "label": "Donor",
       "location": [
@@ -119,11 +115,11 @@
         4.045716762542725
       ],
       "params": {},
-      "width": 397.815673828125
+      "width": 397.8157958984375
     },
     "Frame.001": {
       "bl_idname": "NodeFrame",
-      "height": 378.7057800292969,
+      "height": 373.7057800292969,
       "hide": false,
       "label": "Donor parameters",
       "location": [
@@ -135,7 +131,7 @@
     },
     "Frame.002": {
       "bl_idname": "NodeFrame",
-      "height": 857.6201171875,
+      "height": 855.9534912109375,
       "hide": false,
       "label": "Recipient (Voronoi pattern)",
       "location": [
@@ -147,7 +143,7 @@
     },
     "Frame.003": {
       "bl_idname": "NodeFrame",
-      "height": 330.69970703125,
+      "height": 329.03302001953125,
       "hide": false,
       "label": "Randomize",
       "location": [
@@ -231,8 +227,8 @@
       },
       "width": 140.0
     },
-    "Move.001": {
-      "bl_idname": "SvMoveNodeMk3",
+    "Move": {
+      "bl_idname": "SvMoveNodeMK2",
       "height": 100.0,
       "hide": false,
       "label": "",
@@ -240,9 +236,7 @@
         -254.34125900268555,
         190.09525871276855
       ],
-      "params": {
-        "strength": 1.0
-      },
+      "params": {},
       "width": 140.0
     },
     "Note": {
@@ -550,23 +544,23 @@
     [
       "Plane MK2",
       0,
-      "Move.001",
+      "Move",
       0
     ],
     [
       "Vector Rewire",
       0,
-      "Move.001",
+      "Move",
       1
     ],
     [
       "A Number.001",
       0,
-      "Move.001",
+      "Move",
       2
     ],
     [
-      "Move.001",
+      "Move",
       0,
       "Voronoi 2D",
       0
@@ -609,14 +603,14 @@
     ],
     [
       "Flip normals",
-      "Vertices",
-      "Reroute.001",
+      "Polygons",
+      "Reroute",
       "Input"
     ],
     [
       "Flip normals",
-      "Polygons",
-      "Reroute",
+      "Vertices",
+      "Reroute.001",
       "Input"
     ],
     [
@@ -640,32 +634,14 @@
     [
       "Vector Noise",
       0,
-      "List Math.001",
-      0
-    ],
-    [
-      "Vector Noise",
-      0,
       "List Math",
       0
     ],
     [
       "Vector Noise",
       0,
-      "Map Range",
-      0
-    ],
-    [
-      "List Math",
-      0,
-      "Map Range",
-      1
-    ],
-    [
       "List Math.001",
-      0,
-      "Map Range",
-      2
+      0
     ],
     [
       "Vector Noise",
@@ -690,6 +666,36 @@
       0,
       "Mesh Expression",
       0
+    ],
+    [
+      "Voronoi 2D",
+      0,
+      "Viewer Draw Mk3.001",
+      0
+    ],
+    [
+      "Voronoi 2D",
+      1,
+      "Viewer Draw Mk3.001",
+      1
+    ],
+    [
+      "Vector Noise",
+      0,
+      "Map Range",
+      0
+    ],
+    [
+      "List Math",
+      0,
+      "Map Range",
+      1
+    ],
+    [
+      "List Math.001",
+      0,
+      "Map Range",
+      2
     ],
     [
       "Reroute.001",
@@ -722,6 +728,18 @@
       7
     ],
     [
+      "Move",
+      0,
+      "Viewer Draw Mk3.002",
+      0
+    ],
+    [
+      "Plane MK2",
+      1,
+      "Viewer Draw Mk3.002",
+      1
+    ],
+    [
       "Adaptive Polygons Mk2",
       0,
       "Viewer Draw Mk3",
@@ -732,30 +750,6 @@
       1,
       "Viewer Draw Mk3",
       2
-    ],
-    [
-      "Voronoi 2D",
-      0,
-      "Viewer Draw Mk3.001",
-      0
-    ],
-    [
-      "Voronoi 2D",
-      1,
-      "Viewer Draw Mk3.001",
-      1
-    ],
-    [
-      "Move.001",
-      0,
-      "Viewer Draw Mk3.002",
-      0
-    ],
-    [
-      "Plane MK2",
-      1,
-      "Viewer Draw Mk3.002",
-      1
     ]
   ]
 }
