+++ json_examples/Introduction/BentTorus.json
@@ -2,7 +2,7 @@
   "export_version": "0.079",
   "framed_nodes": {
     "A Number": "Frame.001",
-    "Move.001": "Frame.001",
+    "Move": "Frame.001",
     "Note.001": "Frame",
     "Note.002": "Frame.001",
     "Plane MK2": "Frame.001",
@@ -18,13 +18,11 @@
       "label": "Plane Size",
       "location": [
         -1115.8090147972107,
-        -101.10517883300781
+        -101.10517501831055
       ],
       "params": {
         "float_": 0.0,
-        "float_draft_": 0.0,
         "int_": 6,
-        "int_draft_": 0,
         "selected_mode": "int"
       },
       "width": 140.0
@@ -45,7 +43,7 @@
     },
     "Frame": {
       "bl_idname": "NodeFrame",
-      "height": 547.8262939453125,
+      "height": 547.7862548828125,
       "hide": false,
       "label": "Donor",
       "location": [
@@ -53,11 +51,11 @@
         10.079999923706055
       ],
       "params": {},
-      "width": 333.60284423828125
+      "width": 334.082763671875
     },
     "Frame.001": {
       "bl_idname": "NodeFrame",
-      "height": 524.8558959960938,
+      "height": 525.1359252929688,
       "hide": false,
       "label": "Recipient",
       "location": [
@@ -65,7 +63,7 @@
         -51.75472640991211
       ],
       "params": {},
-      "width": 831.7874145507812
+      "width": 832.2673950195312
     },
     "List Split": {
       "bl_idname": "SvListSplitNode",
@@ -81,8 +79,8 @@
       },
       "width": 140.0
     },
-    "Move.001": {
-      "bl_idname": "SvMoveNodeMk3",
+    "Move": {
+      "bl_idname": "SvMoveNodeMK2",
       "height": 100.0,
       "hide": false,
       "label": "",
@@ -90,9 +88,7 @@
         -484.02160024642944,
         3.855915069580078
       ],
-      "params": {
-        "strength": 1.0
-      },
+      "params": {},
       "width": 140.0
     },
     "Note": {
@@ -205,7 +201,7 @@
       "label": "",
       "location": [
         -675.0626769065857,
-        -49.35852813720703
+        -49.358527183532715
       ],
       "params": {
         "octaves": 2
@@ -295,29 +291,17 @@
     [
       "Plane MK2",
       0,
-      "Move.001",
+      "Move",
       0
     ],
     [
       "Vector Turbulence",
       0,
-      "Move.001",
+      "Move",
       1
     ],
     [
-      "Move.001",
-      0,
-      "List Split",
-      0
-    ],
-    [
-      "A Number",
-      0,
-      "List Split",
-      1
-    ],
-    [
-      "Move.001",
+      "Move",
       0,
       "Viewer Draw Mk3",
       0
@@ -327,6 +311,18 @@
       2,
       "Viewer Draw Mk3",
       2
+    ],
+    [
+      "Move",
+      0,
+      "List Split",
+      0
+    ],
+    [
+      "A Number",
+      0,
+      "List Split",
+      1
     ],
     [
       "Torus",
