+++ json_examples/Introduction/Donut_by_hands.json
@@ -60,8 +60,8 @@
       },
       "width": 140.0
     },
-    "Move.001": {
-      "bl_idname": "SvMoveNodeMk3",
+    "Move": {
+      "bl_idname": "SvMoveNodeMK2",
       "height": 100.0,
       "hide": false,
       "label": "",
@@ -69,9 +69,7 @@
         137.00262451171875,
         245.15399169921875
       ],
-      "params": {
-        "strength": 1.0
-      },
+      "params": {},
       "width": 140.0
     },
     "Number Range": {
@@ -85,10 +83,13 @@
       ],
       "params": {
         "count_": 16,
+        "mode": 3,
         "number_mode": "float",
         "range_mode": "RANGE_COUNT",
         "start_float": 0.0,
+        "step_": 60.0,
         "step_float": 60.0,
+        "stop_": 360.0,
         "stop_float": 360.0
       },
       "width": 140.0
@@ -181,17 +182,17 @@
     [
       "Circle",
       0,
-      "Move.001",
+      "Move",
       0
     ],
     [
       "Vectors in",
       0,
-      "Move.001",
+      "Move",
       1
     ],
     [
-      "Move.001",
+      "Move",
       0,
       "Apply matrix for vectors",
       0
