+++ json_examples/Shapes/Donut.json
@@ -6,8 +6,8 @@
     "Matrix Track To": "Tangent Move",
     "Matrix in": "Tangent Move",
     "Matrix out": "Tangent Move",
-    "Move": "Bottom Flat",
-    "Move.002": "Tangent Move",
+    "Move": "Tangent Move",
+    "Move.001": "Bottom Flat",
     "Note.001": "Tangent Move",
     "Note.003": "Fractal",
     "Note.005": "Tangent Move",
@@ -15,7 +15,7 @@
     "Note.008": "Bottom Flat",
     "Proportional Edit Falloff": "Bottom Flat",
     "Scale": "Fractal",
-    "Scale.002": "Fractal",
+    "Scale.001": "Fractal",
     "Select mesh elements by location.001": "Bottom Flat",
     "Vector Noise": "Fractal",
     "Vector in": "Bottom Flat",
@@ -42,7 +42,7 @@
     },
     "Bottom Flat": {
       "bl_idname": "NodeFrame",
-      "height": 605.8386840820312,
+      "height": 606.2786254882812,
       "hide": false,
       "label": "Make Bottom Flat",
       "location": [
@@ -50,11 +50,11 @@
         10.079999923706055
       ],
       "params": {},
-      "width": 599.3994140625
+      "width": 599.87939453125
     },
     "Fractal": {
       "bl_idname": "NodeFrame",
-      "height": 401.1192626953125,
+      "height": 401.27923583984375,
       "hide": false,
       "label": "Fractal vectors",
       "location": [
@@ -62,7 +62,7 @@
         193.4789276123047
       ],
       "params": {},
-      "width": 606.992431640625
+      "width": 607.472412109375
     },
     "Inset faces": {
       "bl_idname": "SvInsetFaces",
@@ -73,7 +73,6 @@
         2257.6279296875,
         607.4078979492188
       ],
-      "mask_type": [],
       "params": {
         "depth": 0.029999995604157448,
         "inset_type": "region",
@@ -135,21 +134,7 @@
       "width": 140.0
     },
     "Move": {
-      "bl_idname": "SvMoveNodeMk3",
-      "height": 100.0,
-      "hide": false,
-      "label": "",
-      "location": [
-        2974.266345977783,
-        914.8750438690186
-      ],
-      "params": {
-        "strength": 1.0
-      },
-      "width": 140.0
-    },
-    "Move.002": {
-      "bl_idname": "SvMoveNodeMk3",
+      "bl_idname": "SvMoveNodeMK2",
       "height": 100.0,
       "hide": false,
       "label": "",
@@ -158,9 +143,20 @@
         192.87377738952637
       ],
       "params": {
-        "mult_": 1.0,
-        "strength": 1.0
+        "mult_": 1.0
       },
+      "width": 140.0
+    },
+    "Move.001": {
+      "bl_idname": "SvMoveNodeMK2",
+      "height": 100.0,
+      "hide": false,
+      "label": "",
+      "location": [
+        2974.266345977783,
+        910.5752391815186
+      ],
+      "params": {},
       "width": 140.0
     },
     "Note": {
@@ -345,7 +341,7 @@
     },
     "Output": {
       "bl_idname": "NodeFrame",
-      "height": 546.9617919921875,
+      "height": 548.0018310546875,
       "hide": false,
       "label": "Output",
       "location": [
@@ -353,7 +349,7 @@
         10.079999923706055
       ],
       "params": {},
-      "width": 483.674072265625
+      "width": 484.154052734375
     },
     "Proportional Edit Falloff": {
       "bl_idname": "SvProportionalEditNode",
@@ -430,22 +426,7 @@
       "width": 16.0
     },
     "Scale": {
-      "bl_idname": "SvScaleNodeMk3",
-      "height": 100.0,
-      "hide": false,
-      "label": "",
-      "location": [
-        975.0332698822021,
-        832.4366912841797
-      ],
-      "params": {
-        "factor_": 0.1600000262260437,
-        "multiplier": 0.1600000262260437
-      },
-      "width": 140.0
-    },
-    "Scale.002": {
-      "bl_idname": "SvScaleNodeMk3",
+      "bl_idname": "SvScaleNodeMK2",
       "height": 100.0,
       "hide": false,
       "label": "",
@@ -454,8 +435,21 @@
         797.7181854248047
       ],
       "params": {
-        "factor_": 1.0899999141693115,
-        "multiplier": 1.0899999141693115
+        "factor_": 1.0899999141693115
+      },
+      "width": 140.0
+    },
+    "Scale.001": {
+      "bl_idname": "SvScaleNodeMK2",
+      "height": 100.0,
+      "hide": false,
+      "label": "",
+      "location": [
+        975.0332698822021,
+        832.4366912841797
+      ],
+      "params": {
+        "factor_": 0.1600000262260437
       },
       "width": 140.0
     },
@@ -524,7 +518,7 @@
     },
     "Tangent Move": {
       "bl_idname": "NodeFrame",
-      "height": 535.873779296875,
+      "height": 536.5537719726562,
       "hide": false,
       "label": "Move along Tangents",
       "location": [
@@ -532,7 +526,7 @@
         10.079999923706055
       ],
       "params": {},
-      "width": 1703.2313232421875
+      "width": 1703.71142578125
     },
     "Torus": {
       "bl_idname": "SvTorusNode",
@@ -625,7 +619,29 @@
         437.2855052947998
       ],
       "params": {
-        "autosmooth": 1
+        "autosmooth": 1,
+        "lp": [
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0,
+          0
+        ]
       },
       "use_custom_color": true,
       "width": 140.0
@@ -640,9 +656,45 @@
     ],
     [
       "Torus",
+      "Edges",
+      "Reroute.002",
+      "Input"
+    ],
+    [
+      "Torus",
       "Vertices",
       "Reroute",
       "Input"
+    ],
+    [
+      "Reroute",
+      "Output",
+      "Scale",
+      "vertices"
+    ],
+    [
+      "Scale",
+      0,
+      "Vector Noise",
+      0
+    ],
+    [
+      "Vector Noise",
+      0,
+      "Scale.001",
+      0
+    ],
+    [
+      "Scale.001",
+      0,
+      "Vector Rewire",
+      0
+    ],
+    [
+      "Vector Rewire",
+      0,
+      "Matrix in",
+      0
     ],
     [
       "Reroute",
@@ -663,36 +715,6 @@
       3
     ],
     [
-      "Reroute",
-      "Output",
-      "Scale.002",
-      "Vertices"
-    ],
-    [
-      "Scale.002",
-      0,
-      "Vector Noise",
-      0
-    ],
-    [
-      "Vector Noise",
-      0,
-      "Scale",
-      0
-    ],
-    [
-      "Scale",
-      0,
-      "Vector Rewire",
-      0
-    ],
-    [
-      "Vector Rewire",
-      0,
-      "Matrix in",
-      0
-    ],
-    [
       "Matrix Track To",
       0,
       "Matrix Math",
@@ -702,6 +724,24 @@
       "Matrix in",
       0,
       "Matrix Math",
+      1
+    ],
+    [
+      "Matrix Math",
+      0,
+      "Matrix out",
+      0
+    ],
+    [
+      "Reroute",
+      "Output",
+      "Move",
+      "vertices"
+    ],
+    [
+      "Matrix out",
+      0,
+      "Move",
       1
     ],
     [
@@ -723,31 +763,7 @@
       2
     ],
     [
-      "Matrix Math",
-      0,
-      "Matrix out",
-      0
-    ],
-    [
-      "Reroute",
-      "Output",
-      "Move.002",
-      "Vertices"
-    ],
-    [
-      "Matrix out",
-      0,
-      "Move.002",
-      1
-    ],
-    [
-      "Torus",
-      "Edges",
-      "Reroute.002",
-      "Input"
-    ],
-    [
-      "Move.002",
+      "Move",
       0,
       "Inset faces",
       0
@@ -803,31 +819,31 @@
     [
       "Inset faces",
       0,
-      "Move",
+      "Move.001",
       0
     ],
     [
       "Vector in",
       0,
-      "Move",
+      "Move.001",
       1
     ],
     [
       "Proportional Edit Falloff",
       0,
-      "Move",
+      "Move.001",
       2
-    ],
-    [
-      "Move",
-      "Vertices",
-      "Reroute.003",
-      "Input"
     ],
     [
       "Inset faces",
       "Faces",
       "Reroute.004",
+      "Input"
+    ],
+    [
+      "Move.001",
+      "vertices",
+      "Reroute.003",
       "Input"
     ],
     [
